Complete Django Assignment

+++ todoapp/todos/views.py
@@ -1,45 +1,16 @@
-from rest_framework.viewsets import ModelViewSet
-# from rest_framework.response import Response
-# from .models import Todo
-# from .serializers import TodoCreateSerializer
-# from rest_framework.permissions import AllowAny
-# from rest_framework import status
-
-# class TodoAPIViewSet(ModelViewSet):
-#     """
-#         success response for create/update/get
-#         {
-#           "name": "",
-#           "done": true/false,
-#           "date_created": ""
-#         }
-
-#         success response for list
-#         [
-#           {
-#             "name": "",
-#             "done": true/false,
-#             "date_created": ""
-#           }
-#         ]
-#     """
-
-
-
-# views.py
-from rest_framework.permissions import AllowAny
-from rest_framework import viewsets, mixins, generics, status
-from rest_framework.response import Response
-from .models import Todo
-from .serializers import (
-    TodoCreateSerializer,
-    TodoUpdateSerializer,
-    TodoDetailSerializer,
-    TodoListSerializer
-)
-from django.contrib.auth.models import User
-from rest_framework.pagination import PageNumberPagination
+from django.shortcuts import get_object_or_404
+from rest_framework import generics, mixins, status, viewsets
+from rest_framework.authentication import TokenAuthentication
 from rest_framework.decorators import action
+from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import (AllowAny, BasePermission,
+                                        IsAuthenticated)
+from rest_framework.response import Response
+from rest_framework.viewsets import ModelViewSet
+
+from .models import Todo
+from .serializers import (Todo_Create_Serializer, Todo_Detail_Serializer,
+                          Todo_List_Serializer, Todo_Update_Serializer)
 
 
 class TodoPagination(PageNumberPagination):
@@ -47,23 +18,30 @@ class TodoPagination(PageNumberPagination):
     page_size_query_param = 'page_size'
     max_page_size = 50
 
+class IsOwner(BasePermission):
+    
+    def has_object_permission(self,request,view,obj):
+      return obj.user==request.user
 
 class TodoAPIViewSet(viewsets.ViewSet):
     """
     Handles Create, Update, Get, Delete and List operations for Todo.
     """
-    permission_classes = [AllowAny]
+    permission_classes =    [IsAuthenticated & IsOwner]
+    authentication_classes= [TokenAuthentication]
+    # permission_classes=     [AllowAny]
     pagination_class = TodoPagination
 
 
     def create(self, request):
-        serializer = TodoCreateSerializer(data=request.data)
+        serializer = Todo_Create_Serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         todo = serializer.save()
         return Response({
             "todo": todo.name,
             "done": todo.done,
             "Date_created": todo.date_created,
+            'id':todo.id
         })
        
     @action(detail=False, methods=['put'], url_path='update')
@@ -72,8 +50,8 @@ class TodoAPIViewSet(viewsets.ViewSet):
             todo = Todo.objects.get(id=request.data.get("todo_id"))
         except Todo.DoesNotExist:
             return Response({"error": "Todo not found"}, status=404)
-
-        serializer = TodoUpdateSerializer(todo, data=request.data)
+        self.check_object_permissions(request, todo)
+        serializer = Todo_Update_Serializer(todo, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
@@ -84,12 +62,14 @@ class TodoAPIViewSet(viewsets.ViewSet):
             todo = Todo.objects.get(pk=pk)
         except Todo.DoesNotExist:
             return Response({"error": "Todo not found"}, status=404)
-        serializer = TodoDetailSerializer(todo)
+        self.check_object_permissions(request, todo)
+        serializer = Todo_Detail_Serializer(todo)
         return Response(serializer.data)
 
     def destroy(self, request, pk=None):
         try:
             todo = Todo.objects.get(pk=pk)
+            self.check_object_permissions(request, todo)
             todo.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except Todo.DoesNotExist:
@@ -101,11 +81,8 @@ class TodoAPIViewSet(viewsets.ViewSet):
         if not user_id:
             return Response({"error": "user_id is required"}, status=400)
         todos = Todo.objects.filter(user_id=user_id).order_by('-date_created')
+        self.check_object_permissions(request, todos)
         paginator = self.pagination_class()
         paginated_qs = paginator.paginate_queryset(todos, request)
-        serializer = TodoListSerializer(paginated_qs, many=True)
+        serializer = Todo_List_Serializer(paginated_qs, many=True)
         return paginator.get_paginated_response(serializer.data)
-
-
-
-
